major updates made.

+++ app/Http/Controllers/OrderController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 use App\Order;
-use App\Warehouse;
+use App\WarehouseItem;
 use DB;
 use Illuminate\Http\Request;
 
@@ -38,13 +38,25 @@ class OrderController extends Controller
       */
     public function store(Request $request)
     {
+      $id = $request->input('warehouseitem_id');
+      $quantityPurchased = $request->input('numberofitems');
+         
+      $warehouseitem = WarehouseItem::find($id);
+      if (empty($warehouseitem)){echo "No results For id " . $id; return;}
+      $previousQuantity= $warehouseitem->quantity;
+      $newQuantity= $previousQuantity + $quantityPurchased;     
+      $warehouseitem->quantity = $newQuantity;
+      $warehouseitem->save();
+      
+      echo "warehouse item quantity has been Updated. /n";  
+        
       $order = new Order;
       $order->issuingofficername = $request->input('issuingofficername');
       $order->supplycompanyname = $request->input('supplycompanyname');
       $order->supplyofficername = $request->input('supplyofficername');
-      $order->warehouse = $request->input('warehouse');
-      $order->numberofcartons = $request->input('numberofcartons');
-      $order->quantityliters = $request->input('quantityliters');
+      $order->warehouseitem_id = $request->input('warehouseitem_id');
+      $order->numberofitems = $request->input('numberofitems');
+      $order->quantityofliters = $request->input('quantityofliters');
       $order->priceperitem = $request->input('priceperitem');
       $order->total = $request->input('total');
       $order->notes = $request->input('notes');
@@ -66,8 +78,8 @@ class OrderController extends Controller
       $order->supplycompanyname = $request->input('supplycompanyname');
       $order->supplyofficername = $request->input('supplyofficername');
       $order->warehouse = $request->input('warehouse');
-      $order->numberofcartons = $request->input('numberofcartons');
-      $order->quantityliters = $request->input('quantityliters');
+      $order->numberofitems = $request->input('numberofitems');
+      $order->quantityofliters = $request->input('quantityofliters');
       $order->priceperitem = $request->input('priceperitem');
       $order->total = $request->input('total');
       $order->notes = $request->input('notes');
@@ -82,9 +94,20 @@ class OrderController extends Controller
       *@return Response
       */
     public function destroy($id)
-    {
+    {   
       $order = Order::find($id);
       if (empty($order)){echo "No results For Matching"; return;}
+      
+      $warehouseID= $order->warehouseitem_id;
+    
+      $warehouseitem = WarehouseItem::find($warehouseID);
+      if (empty($warehouseitem)){echo "No results For id " . $id; return;} 
+      $purchasedQuantity = $order->numberofitems;
+      $existingQuantity = $warehouseitem->quantity;    
+      $newQuantity = $existingQuantity - $purchasedQuantity;
+      $warehouseitem->quantity = $newQuantity; 
+      $warehouseitem->save();  
+        
       $order->delete();
       echo "Item has been Deleted.";
       return;
